property-based-testing/pytest-quickcheck/requirements.txt
pytest-quickcheck==0.8.3